move mapper tests to crate tests module
